requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0
anthropic==0.15.0
openai==1.10.0
sqlalchemy==2.0.25
python-dotenv==1.0.1
sqlparse==0.4.4
pandas==2.2.0
tiktoken==0.5.2
httpx==0.26.0
python-multipart==0.0.6
sse-starlette==2.0.0
